requirements.txt
APScheduler==3.9.1
cachetools==5.0.0
certifi==2021.10.8
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.1
